requirements.txt
telethon==1.29.0
Pillow==9.2.0
boto3==1.26.0
pyyaml
